Refac: rename protocol
