+++ src/careamics/models/constraints/__init__.py
@@ -2,6 +2,6 @@
 
 __all__ = ["ModelConstraints", "UNetConstraints", "get_model_constraints"]
 
+from .model_constraints import ModelConstraints
 from .model_constraints_factory import get_model_constraints
-from .model_constraints_protocol import ModelConstraints
 from .unet_constraints import UNetConstraints
